GH #57: Builder: use Installer to unpack existing parcels:
When the builder finds a requirement that is already packaged,
it will unpack it in the build directory. Unfortunately this was
a quick hack and was not recursive.

GH #57 shows this had an effect of inner dependencies being
missing and breaking the build.

Since we now have a proper installer that can install separately
from the set-up process of the install process, we can use that
directly.

This can still be smoothed over in the future:

* We need to pass to the installer the directory straight from
  the "bundler_args" which is sub-optimal.
* We need to still verify that the parcel file exists in the
  builder and only then call the installer. Also sub-optimal.
* The builder and installer specify the packages in different
  strings and we need to translate from one to the other. Very
  simple, but will be unnecessary if we had package objects.