Improvements to build cmd heading to new Requirement object:
This luckily doesn't break anything yet and is a relatively small
change.

We introduce cleanups and improvements to the "build" CLI command
in order to accommodate the upcoming Requirement object:

* Introduce a new constant: PAKKTE_PACKAGE_SPEC.

The spec provides a way to indicate a required package ("perl/Plack")
or a required package with a version ("perl/Plack=0.008"). It does
not support ranges and the syntax is for any kind of package, no
matter what category and what versioning scheme it supports.

This means that you cannot use a version that is equivalent in
meaning, only in exact characters.

* Add requirements on some parameters in getopt.

* Add "--from-index" option so we could just build from the index
  instead of providing "--json-file". You may still provide it, but
  you can now just provide the index and say "--from-index" to note
  you want to build everything in it.