Extract package archives with correct names
When we donwload a package archive, such as Getopt-Long-v2.49.1.tar.gz,
the top-level directory inside this archive may have a name that does
not exactly match the version spec for that package; in this example,
the top-level directory is called Getopt-Long-2.49.1/ (notice the 'v'
character is missing). This of course causes a missmatch between the
package configuration and the directory where the package is extracted.

To fix this, we extract the package into a temporary directory and
handle three possible cases:

1. The package extracts into a single top-level directory: use that
   directory as the output, but with a canonical name.
2. The package extracts into '.': put all the contents of the package
   into a directory with a canonical name.
3. The package contains no files: warn and skip.

In all cases, the canonical name for a package directory is a name based
on the package name and version used in the corresponding TOML file.