Handle conflicts between PERL5LIB, @INC, and maybe sitecustomized.pl:
perl/Encode would not compile because it needs a 'Symbol.pm' in a
build directory it creates but accidentally loaded the Symbol core
perl module.

The reason is that while we add '.' into PERL5LIB (to avoid it being
removed from @INC in our local sitecustomized.pl), it represents the
directory in which it chdir'ed instead of where the build started.

Anyway, confusing, very confusing, and now it's cleaned up by taking
the actual directory where we started a build.