Added bootstrapping support
Several modules need to be built before the others and without
their recursive dependencies.
This allows us to resolve some cyclic dependencies in the most
basic toolchain modules.

For now the list is hardcoded and kept to the minimal list of
modules that were necessary to have bootstrapped.
We can later think of ways to improve this.