Untangle the Scaffolder from the TOML format and Config:
The Scaffolder was the only user of the ConfigReader and
ConfigReader::TOML. It has now moved to Config::Any, which I think
we will end up using all over, not just for the Scaffolder.

Moving this from the Scaffolder namespace will probably be a future
step in having a general configruation for everything.