installer: use requested by user versions to resolve dependencies
This fix a bug when user wants to install package X=1
but there are dependencies which try to install X
and installer for dependencies is using latest version of X in spec repo,
which may be X=2 and throws an error:

perl/Moo=2.003002 already installed. Cannot install new version:2.002004

Now we will use requested by user versions if it exists.