Enable non atomic installs
Pakket currently uses a simple and effective system to install all
dependencies atomically. It involves copying the entire installation
directory to a new working directory, and later changing symlinks.

The issue with that is that with Docker images, it's really wasteful. It
creates a new layer with the entire size of the installation, even if a
single small module was installed or upgraded.

This patch creates a configuration option, `atomic`, and a command-line
option for `pakket install`, `--no-atomic`, which control this
behaviour.

By default, nothing changes, so atomic installations are always on
unless explicitly requested with:

In /etc/pakket.json:

```json
{
    ...
    "atomic": 0
}
```

Or:

```
$ pakket install --no-atomic ...
```